Address GroundCraft review feedback
